separated test from solution
