+++ src/main/scala/ru/spbau/jvm/scala/task169/Solution.scala
@@ -28,22 +28,10 @@ sealed trait Nat {
   type Flip <: Nat
   type DivBy2 <: Pair
   type Mult[C <: Nat] <: Nat
-  // Base ^ this
-  type Deg[Base <: Nat] <: Nat
   type F <: Nat
   type F0 <: Nat // no 1, all numbers are 2 or greater
   type F1 <: Nat // one 1
   type F11 <: Nat // two 1
-}
-
-sealed trait Pair {
-  type First <: Nat
-  type Second <: Nat
-}
-
-class ParametrizedPair[A <: Nat, B <: Nat] extends Pair {
-  type First = A
-  type Second = B
 }
 
 class O extends Nat {
@@ -53,7 +41,6 @@ class O extends Nat {
   type DivBy2 = ParametrizedPair[O, O]
   type Flip = S[O]
   type Mult[_ <: Nat] = O
-  type Deg[_ <: Nat] = S[O]
   type F = S[O]
   type F0 = S[O]
   type F1 = O
@@ -68,11 +55,20 @@ class S[T <: Nat] extends Nat {
   type Flip = O
   type DivBy2 = ParametrizedPair[T#DivBy2#First#Plus[T#DivBy2#Second], T#DivBy2#Second#Flip]
   type Mult[C <: Nat] = T#Mult[C]#Plus[C]
-  type Deg[Base <: Nat] = T#Deg[Base]#Mult[Base]
   type F = F0#Plus[F1#Plus[F11]]
   type F0 = DivBy2#Second#IsZero[DivBy2#First#F, O, Nat]
   type F1 = T#F0
   type F11 = T#F1
+}
+
+sealed trait Pair {
+  type First <: Nat
+  type Second <: Nat
+}
+
+class ParametrizedPair[A <: Nat, B <: Nat] extends Pair {
+  type First = A
+  type Second = B
 }
 
 object Solution {
@@ -90,25 +86,4 @@ object Solution {
   implicit def repS[T <: Nat](implicit rep: NatRep[T]) = new NatRep[S[T]](rep.i + 1)
 
   def toInt[T <: Nat](implicit rep: NatRep[T]) = rep.i
-
-  def solve(i : Int): Int = {
-    if (i == 0) {
-      return 1
-    }
-    if (i % 2 == 1) {
-      return solve(i / 2)
-    }
-    solve(i / 2) + solve(i / 2 - 1)
-  }
-
-  def main(args: Array[String]): Unit = {
-    assert(toInt[_1#F] == solve(1))
-    assert(toInt[_2#F] == solve(2))
-    assert(toInt[_3#F] == solve(3))
-    assert(toInt[_4#F] == solve(4))
-    assert(toInt[_10#F] == solve(10))
-    assert(toInt[_10#F] == 5)
-    assert(toInt[_25#F] == solve(25))
-    assert(toInt[_100#F] == solve(100))
-  }
 }
